Update README.md
The purpose of this assignment is to warm-up with Scala and to refresh preliminaries from prior courses.

Learning Goals. The primary learning goals of this lab are to build intuition for the following:

thinking in terms of types, values, and expressions;

representing data structures as algebraic data types; and

representing programs as abstract syntax trees.

PL Ideas Abstract syntax. Expressions conforming to types and evaluating to values.

FP Skills Data structures as algebraic terms.